Fix handler so it binds properly with sockets, and rename socket server file due to weird babel issue
