+++ test/server/socket/events.test.js
@@ -1,8 +1,15 @@
 import test from 'ava';
-import {match} from 'sinon';
+import proxyquire from 'proxyquire';
+import {stub, match} from 'sinon';
 
-import eventsArray from '../../../server/socket/events';
 import {MockSocket} from '../mocks/socket';
+import {repositories} from '../mocks';
+import Player from '../../../server/models/player';
+
+const eventsArray = proxyquire('../../../server/socket/events', {
+  '../repositories': repositories
+}).default;
+const {playerRepository} = repositories;
 
 // Remap events to object for ease of use
 const events = {};
@@ -17,4 +24,17 @@ test('"initPlayer" creates a new player if none exists', t => {
 
   t.not(socket.playerId, null);
   t.true(socket.emit.calledWith('setPlayerId', match.string));
+});
+
+test('"initPlayer" retrieves an existing player if the player is reconnecting', t => {
+  const originalSocket = new MockSocket();
+  const player = new Player(originalSocket);
+  playerRepository.find = stub().withArgs(player.id).returns(player);
+
+  const socket = new MockSocket();
+
+  events.initPlayer(socket);
+
+  t.is(socket.playerId, player.id);
+  t.true(socket.emit.calledWith('setPlayerId', player.id));
 });
